Try and light up 59 non-faulty drives
This is for the situation where there's >1 LUN missing in the 0-59
range, so we can't just say "pull that one", *and* the /dev/ entry for
the failed drive is missing so we can't just say "pull the one with
the light on". Instead, we light up all the drives we have /dev
entries for on controller 0, which should hopefully be the 59 good
drives.

perl is used because you can't do the newline -> space conversion in
sed, and xargs can't cope with putting its stuff in the middle of a
command-line (-I only substitutes on value in).