Always light up non-failed drives if disk is missing
Previously, if find_missing_disk_bay was able to determine which disk
bay corresponded to the failed disk, we didn't bother lighting up the
non-failed drives.

On reflection, it's better to do this anyway, as a visual aid to the
person doing the disk swap.

Also, there was a logic error where if a single bay _was_ found, then
you go an error message about failing to light up non-faulty drives.